requirements.txt
asgiref==3.8.1
Django==3.2.16
djangorestframework==3.12.4
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
